+++ src/Sentry/Protocol/Metrics/SetMetric.cs
@@ -1,0 +1,15 @@
+using Sentry.Extensibility;
+using Sentry.Internal.Extensions;
+
+namespace Sentry.Protocol.Metrics;
+
+/// <summary>
+/// Sets track a set of values on which you can perform aggregations such as count_unique.
+/// </summary>
+internal class SetMetric : Metric
+{
+    private HashSet<int> Value { get; set; } = new();
+
+    public override void WriteConcreteProperties(Utf8JsonWriter writer, IDiagnosticLogger? logger) =>
+        writer.WriteArrayIfNotEmpty("value", Value, logger);
+}
